requirements.txt
fastapi==0.134.0
uvicorn==0.41.0
livekit==1.1.2
livekit-api==1.1.0
httpx==0.28.1
numpy==2.4.2
python-dotenv==1.2.1
PyJWT==2.11.0